% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{create_project_input}
\alias{create_project_input}
\title{interface: interface visualization for imageTCGA}
\usage{
create_project_input(db)
}
\description{
interface: interface visualization for imageTCGA
}
