% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputs.R
\name{generate_summary_statistics}
\alias{generate_summary_statistics}
\title{outputs: output visualization function for imageTCGA}
\usage{
generate_summary_statistics(filtered_data)
}
\description{
outputs: output visualization function for imageTCGA
}
