% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_new.R
\name{imageTCGAnew}
\alias{imageTCGAnew}
\title{imageTCGAnew: A Shiny application to explore the TCGA DiagnosticImageDatabase}
\usage{
imageTCGAnew()
}
\description{
imageTCGAnew: A Shiny application to explore the TCGA DiagnosticImageDatabase
}
