% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{imageTCGA}
\alias{imageTCGA}
\title{imageTCGA: A Shiny application to explore the TCGA Diagnostic Image Database}
\usage{
imageTCGA()
}
\description{
imageTCGA: A Shiny application to explore the TCGA Diagnostic Image Database
}
