% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imageTCGA.R
\name{imageTCGA}
\alias{imageTCGA}
\title{imageTCGA: A Shiny application to explore the TCGA Diagnostic Image Database}
\usage{
imageTCGA()
}
\value{
A Shiny application object
}
\description{
This function launches a Shiny application that allows users to explore and
download data from the TCGA Diagnostic Image Database.
}
\examples{
if (interactive()) {
    BiocHubsShiny()
}

}
